[Java.Interop{,.Dynamic,.Export}] Set Version to 0.1.0.0 (#894)
Commit 9841c214b2dc45033a87b6548a4256db3a4b94cd
Context: https://discord.com/channels/732297728826277939/732297837953679412/898675755779768330

An unfortunate unexpected breakage occurred due to commit 9841c214:
the version of `Java.Interop.dll` changed -- which we *should* have
expected but completely overlooked & forgot -- which in turn meant
that *everything* which depends upon it -- which is everything --
started breaking due to the version change, e.g.

	error CS1705: Assembly 'Microsoft.Maui' with identity 'Microsoft.Maui, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null'
	uses 'Java.Interop, Version=0.1.6.0, Culture=neutral, PublicKeyToken=84e04ff9cfb79065'
	which has a higher version than referenced assembly
	'Java.Interop' with identity 'Java.Interop, Version=0.1.2.0, Culture=neutral, PublicKeyToken=84e04ff9cfb79065'

As the primary motivation with 9841c214 was to improve the use of
*MSBuild* task assemblies -- which `Java.Interop.dll` is *not* --
modify the logic of 9841c214 so that:

 1. `build-tools/scripts/Version.props` allows the `$(Version)`
    property to be *overridden*, and

 2. Override `$(Version)` within `src/Java.Interop`,
    `src/Java.Interop.Dynamic`, and `src/Java.Interop.Export`.

`Java.Interop.Dynamic.dll` and `Java.Interop.Export.dll` aren't
shipped (yet), but we would like them to eventually, and it's "nicer"
if they all share the same assembly version.