[One .NET] 'dotnet publish' should support $(AndroidPackageFormats) (#6403)
Fixes: https://github.com/xamarin/xamarin-android/issues/6393

We found that `dotnet build -c Release` doesn't copy `.apk` files to
`$(PublishDir)`.

In f9f879c7, `$(AndroidPackageFormats)` was changed to default to
`aab;apk`. However, the `dotnet publish` target was not updated for
this change.

I updated an existing test, and made changes so all
`$(AndroidPackageFormats)` files will be copied.