[build] Some fixes needed for `dotnet build Xamarin.Android.sln` (#6390)
Some fixes that were required to allow `Xamarin.Android.sln` to build
with `dotnet build` on macOS:

	dotnet build Xamarin.Android.sln -m:1

Largely this is just path separator changes for instances when
`dotnet build` doesn't change the path separator where Mono's
`msbuild` will; consider the following `<Exec/>` invocation:

	<Exec Command="&quot;$(JarPath)&quot; uf &quot;$(_MonoAndroidJar)&quot; -C &quot;$(IntermediateOutputPath)jcw\bin&quot; ." />

On Mono's `msbuild`, `$(IntermediateOutputPath)jcw\bin` will be
treated as `$(IntermediateOutputPath)jcw/bin` on non-Windows
platforms.  `dotnet build` does not perform this "normalization".

Additionally, change `api-compat` to run on `dotnet` when building
with `dotnet build`.  Mono allows you to `Process.Start ("blah.exe")`
and it will automagically change it to `"mono blah.exe"`.
.NET doesn't do this, so we need to explicitly call `mono` or
`dotnet` when running on .NET.  We went with `dotnet` as long as we
were making changes anyways.

Note: This is not *everything* required in order to support
`dotnet build Xamarin.Android.sln` on macOS.  It's gets us closer,
particularly with 572ede52, but some issues remain.

Note: This does not attempt to eliminate use of Mono from the build;
MSBuild targets which use the `<Exec>` task to run `.exe` processes
such as `api-merge.exe` continue to use Mono.

Co-authored-by: Jonathan Peppers <jonathan.peppers@microsoft.com>