[tests] Use `emulator -logcat-output` to capture device state (#6432)
Sometimes unit tests will fail, because the emulator failed to start:

	Task "StartAndroidEmulator"
	  …
	  Tool /Users/runner/Library/Android/sdk/emulator/emulator execution started with arguments:  -verbose -no-boot-anim -no-audio -no-snapshot -cache-size 512 -timezone "Etc/UTC" -avd XamarinAndroidTestRunner29-x86_64 -port 5570
	  …
	  [emulator stderr] handleCpuAcceleration: feature check for hvf
	…
	Task "Xamarin.Android.Tools.BootstrapTasks.Adb"
	  …
	  Executing: /Users/runner/Library/Android/sdk/platform-tools/adb -s emulator-5570 wait-for-device
	  …
	/Users/runner/work/1/s/build-tools/scripts/TestApks.targets(78,5): error :   Command /Users/runner/Library/Android/sdk/platform-tools/adb -s emulator-5570 wait-for-device failed with exit code -1 [/Users/runner/work/1/s/tests/Mono.Android-Tests/Mono.Android-Tests.csproj]
	Done executing task "Xamarin.Android.Tools.BootstrapTasks.Adb" -- FAILED.
	Done building target "AcquireAndroidTarget" in project "Mono.Android-Tests.csproj" -- FAILED.

What's happening is that `adb wait-for-device` is timing out after
2 minutes, at which point we give up.

We don't know *why* `adb wait-for-device` is timing out.  We have no
additional information to investigate.

Update the `<StartAndroidEmulator/>` task to have a
`StartAndroidEmulator.LogcatFile` property, which is used with the
`emulator -logcat-output` option.  This allows us to capture *all*
output that `adb logcat` would provide, including during emulator
startup.  This should allow us to further diagnose *why* the emulator
is "hanging" during `adb wait-for-exit`.

Store the full `adb logcat` output into the file
`bin/Test$(Configuration)/logcat-$(Configuration)-full.log`.
This file will be included in the test `xa-build-status*` zip
archives which are saved at the end of a test run.

Additionally, fix a bug in the `DeviceTest.RestartDevice()` method
where it could fail when using `dotnet`:

	RunProcess: /Users/runner/Library/Android/dotnet/dotnet /Users/runner/work/1/s/bin/TestRelease/Emulator.csproj /restore /t:AcquireAndroidTarget
	(1, , Could not execute because the specified command or file was not found.

We need to be using `dotnet build /path/to/proj.csproj`.