[ci] Improve build and test result packaging (#6411)
Attempt to improve our ability to analyze build and test failures with
a handful of changes.

The Xamarin.Android.Build.Tests and MSBuildDeviceIntegration test suites
will now generate and build test projects in a [temporary directory][0]
when running on CI.  A timestamped directory has been appended to this
new test output path, to serve as an extra precaution for identifying
any potentially old results.  However, the docs mention that the output
folder should always be clean at the start of each job:

    Note: Build.ArtifactStagingDirectory and Build.StagingDirectory are
    interchangeable. This directory is purged before each new build, so
    you don't have to clean it up yourself.

The result packaging MSBuild targets have been replaced with a step in
the xaprepare tool.  We will now upload an unzipped directory structure
that can be expanded and viewed in the pipelines artifact tab.  This
should make it easier to download an individual log file for a failing
test without having to download zip (or in some cases nested zip) files.

Finally, we will now only upload results when a job has a failing step.
This should make it easier to quickly find the artifact that corresponds
to a failing job, as we will no longer have to search through dozens of
similarly named artifacts for a single test failure.

[0]: https://docs.microsoft.com/en-us/azure/devops/pipelines/build/variables?view=azure-devops&tabs=yaml#build-variables-devops-services