[illink] fix XA2000 warning, again
Context: https://github.com/dotnet/linker/issues/2333

In testing 10b97206, I found that the XA2000 warning didn't actually
appear when consuming a class library with `AppDomain` usage in .NET 6.

After some discussion on dotnet/linker#2333, the only way to log a
warning from a custom trimmer step is:

    Context.LogMessage (MessageContainer.CreateCustomWarningMessage (Context, msg, 6200, new MessageOrigin (), WarnVersion.ILLink5))

Where the code must be 6001 or higher, and it will automatically be
prefixed with `IL`. I mapped `XA2000`, to `IL6200`, to just have the
number *similar*? We can chooose something else, let me know.

Once the warning was working, I got *many* warnings for the same
assembly. It turns out we were emitting it in .NET 6 for every type!

    protected void ProcessType (TypeDefinition type)
    {
        var assembly = type.Module.Assembly;
        if (!CheckShouldProcessAssembly (assembly))
            return;
    //...
    bool CheckShouldProcessAssembly (AssemblyDefinition assembly)
    {
        if (!Annotations.HasAction (assembly))
            Annotations.SetAction (assembly, AssemblyAction.Skip);
        if (IsProductOrSdkAssembly (assembly))
            return false;
        // Called for each type!
        CheckAppDomainUsage (assembly, (string msg) =>

To solve this issue, I created a new `HashSet<string>` of assembly
names. This way we will only log this warning once per assembly.

I updated a test for this scenario.