[generator] emit UnconditionalSuppressMessage in __TypeRegistrations (#908)
Building a .NET MAUI app with
`-p:SuppressTrimAnalysisWarnings=false -p:TrimmerSingleWarn=false`
shows the warning:

	src\Compatibility\Android.FormsViewGroup\src\obj\Release\net6.0-android\generated\src\Java.Interop.__TypeRegistrations.cs(35,4):
	error IL2057: Java.Interop.__TypeRegistrations.Lookup(String[],String): Unrecognized value passed to the parameter 'typeName' of method 'System.Type.GetType(String)'.
	It's not possible to guarantee the availability of the target type.

This code path should not even be used normally.  It is only called
when running inside the Android Designer.

For now, let's simply emit an extra attribute to ignore the warning:

	[System.Diagnostics.CodeAnalysis.UnconditionalSuppressMessage("Trimming", "IL2057")]

This attribute was added in .NET 5, so we can always emit the code
and use `#if NET5_0_OR_GREATER`.