[ci] Add pipeline for updating api docs (#6460)
Add a new `azure-pipelines-apidocs.yaml` pipeline that can be used to
update the `Mono.Android` API docs in the
[xamarin/android-api-docs][0] repository.  We will now only build
`src/Mono.Android` with generated JavaDoc XML when running this new
pipeline.

The new pipeline will run the `<UpdateExternalDocumentation/>` target
and upload the contents of `/android-api-docs/docs/Mono.Android/en/`
as a build artifact.  These updated docs can be downloaded for manual
review and pull request creation.

The .NET 6 packs have been updated to include the same
`Mono.Android.xml` docs that are bundled into the traditional Windows
`.vsix` installer.  These docs are generated from the content of
`external/android-api-docs` when the `<AssembleApiDocs/>` installer
target is run.

[0]: https://github.com/xamarin/android-api-docs