[One .NET] Add non-spamming warning IL6200, née XA2000 (#6449)
Context: https://github.com/dotnet/linker/issues/2333

In testing 10b97206, I found that the XA2000 warning didn't actually
appear when consuming a class library with `AppDomain` usage in a
.NET 6 SDK for Android app.

After some discussion on dotnet/linker#2333, the only way to log a
warning from a custom trimmer step is:

	Context.LogMessage (
	    MessageContainer.CreateCustomWarningMessage (
	      Context, msg, 6200, new MessageOrigin (), WarnVersion.ILLink5))

Where the code must be 6001 or higher, and it will automatically be
prefixed with `IL`.  I mapped `XA2000`, to `IL6200`, to just have a
*similar* value.

Once the warning was working, I got *many* warnings for the same
assembly.  It turns out that, due to how
`FixAbstractMethodStep.ProcessType()` and
`FixAbstractMethodStep.CheckShouldProcessAssembly()` interacted,
if *any* method called `AppDomain.CreateDomain()`, then *every type*
in the assembly would elicit a warning for the assembly:

	partial class FixAbstractMethodStep {

	    // ProcessType() is called once per type in a given library
	    protected void ProcessType (TypeDefinition type)
	    {
	        var assembly = type.Module.Assembly;

		// We thus call CheckShouldProcessAssembly() once per type
	        if (!CheckShouldProcessAssembly (assembly))
	            return;
	        // …
	    }

	    bool CheckShouldProcessAssembly (AssemblyDefinition assembly)
	    {
	        // …
	        // We thus warn for each separate type!
	        CheckAppDomainUsage (assembly, (string msg) =>
	    }
	}

This would spam the build log; our [unit test][0] contains only one
call to `AppDomain.CreateDomain()`, but would result in *28* warnings,
as the library contained 7 types, and was built across 4 RIDs.

To solve this issue, I created a new `HashSet<string>` of assembly
names.  This way we will only log this warning once per assembly.

I updated a test for this scenario.

[0]: https://github.com/xamarin/xamarin-android/blob/85a50c8099731971f8dd7c927e683e83f5462ea8/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Tasks/LinkerTests.cs#L181-L211