[Mono.Android] Pass frameworks.xml file to mdoc
A `frameworks.xml` file will now be passed to `mdoc.exe` when updating
external API docs.  This file contains the metadata required to find
and update or create [FrameworksIndex files][0].

The version of `Mono.Android.dll` used to generate docs will no longer
be passed to `mdoc.exe`.  Instead, it will be copied to the relative
`%(Source)` directory described in `frameworks.xml` to be resolved
automatically.

[0]: https://github.com/xamarin/android-api-docs/tree/cd414a93057b6f61afcdd3db6ff44391fb166246/docs/Mono.Android/en/FrameworksIndex