Don't package empty AOT-d DSOs
Fixes: https://github.com/xamarin/xamarin-android/issues/3932

Profiled AOT tends to produce a number of "empty" .so shared libraries
because the profile names only a handful of types from a handful of
assemblies and the AOT compiler in Mono currently doesn't support not
outputting the .so in such cases.

This has the unfortunate effect that an application may have a large
number of such empty .so libraries and each of them will have to be
opened and loaded before Mono can determine that there's no code in them
and the JIT has to compile the requested types and methods.  Loading
every .so takes a non-zero amount of time, especially if the shared
libraries aren't unpacked to the filesystem.  I've seen times between 10
and 30ms to load a single shared library, which can add up to a
non-trivial amount of time wasted loading empty libraries.

This commit adds code to examine and ignore such .so files which don't
include any executable code.  The checks are crafted so that only AOT-d
libraries are considered for rejection.

Time savings are around 3ms for a plain XA app and around 10ms for a
sample MAUI app on the Pixel 3 XL phone.  In the latter case 49 shared
libraries out of 120 total are rejected and not packaged, contributing
to (small, each empty DSO is between 7 and 25k) space savings in the APK
in addition to slightly better startup performance.