[Mono.Android] Update F# API docs (#6487)
Currently, when import & update Android API docs (cee5e727, 0503d636),
the order of the `<TypeSignature/>` element is changed somewhat
randomly.  This creates "noise" in the xamarin/android-api-docs diff.

WHile this only happens on the first time the
`UpdateExternalDocumentation` target is run, it's not ideal.

Update the `_RunMdoc` target -- called by the
`UpdateExternalDocumentation` target -- so that
`mdoc update --lang fsharp` is also used.  This ensures consistent
ordering of the `<TypeSignature/>` element, and also ensures that the
F# member information appears *after* C# member information.