[.NET] Stop .aar files being picked up from sub projects.
PR #6485 reworked the HelloWorld samples example to fix the following
issue.

```
error APT2144: invalid file path '/Users/.../Documents/Sandbox/Xamarin/designerrewrite/samples/HelloWorld/obj/Debug/net6.0-android/lp/1.stamp'
```

This was occurring because when we built the `HelloWorld` app it would
automatically pickup various aar files in the `HelloWorldLibrary` project.
This project was  in a sub directory of the `HelloWorld` app.

What we would end up with is the app would end up including both
`Debug` and `Release` .aar files from the sub directory (if both were built).
But in addition it would also run a search for other `.aar` files which
would then be added with the `AndroidSkipResourceProcessing` metadata
set to `false`. So we would end up with data such as

```
AndroidAarLibrary=HelloWorldLibrary/bin/Debug/HelloWorldLibrary.aar
            AndroidSkipResourceProcessing=true
            Pack=true
AndroidAarLibrary=HelloWorldLibrary/bin/Release/HelloWorldLibrary.aar
            AndroidSkipResourceProcessing=true
            Pack=true
AndroidAarLibrary=/Users/foo/path/to/xa/samples/HelloWorldLibrary/bin/Debug/HelloWorldLibrary.aar
            AndroidSkipResourceProcessing=false
            Pack=false
```

These are ALL the same library and they should only be included once.
So we have a few fixes for this. The first is to alter the `AutoImport`
code to ignore sub directories depending on the build `$(Configuration)`.
This is so when building `Debug` we do NOT include `Release` in the
`AutoImport`.

Next is to alter the `_ResolveAars` target to exclude the auto imported
items from the `.aar` libraries we find during the search. This has to
be done using the `FullPath` metadata as the search will result in paths
which are full paths, but auto import produces relative ones. This is why
the old `Exclude` code was not working as expected.