Don't count assemblies to be skipped when packaging
Fixes: https://developercommunity.visualstudio.com/t/Xamarin-Forms-Android-App-crashes-on-sta/1578587
Context: c9270261

c9270261 added support for assembly stores which, in turn, depend on
knowing the exact number of assemblies packaged in the APK in order to
properly perform binary search on the tables of assembly name hashes.
This information is recorded by the `GeneratePackageManagerJava` task
in the `libxamarin-app.so` library on the build time and then used by
the assembly store code on the run time.

The code responsible for counting failed to skip assemblies which had
the `AndroidSkipAddToPackage` metadata, in our case the old Android
Support Library ones, replaced by AndroidX, with the metadata set on the
old assemblies by the AndroidX migration code.

Properly ignore skipped assemblies when counting.

Additionally, improve the `assembly-store-reader` utility to show the
hash tables stored in the blob as well as check whether they are
properly sorted.