[monodroid] Fix the `timing=bare` logging category (#6499)
Commit 917e1a94 rewrote `init_logging_categories()`, which parses the
`debug.mono.log` system property, to eliminate memory allocation and
perform the parsing faster.  However, the conversion failed to set
the `LOG_TIMING` category when the `timing=bare` is used, which
results in the timing information not showing in the logs.

Fix the mistake by setting `LOG_TIMING` where appropriate.