[monodroid] Fix `mono_log_mask` parsing and add property docs (#6494)
From the "how did I miss it?!" department.

The main purpose of this commit is to fix parsing of the
`mono_log_mask` component of the `debug.mono.log` property.
The component is set to a list of Mono logging categories (e.g. `asm`,
`gc` or `dll`) which, so far, was supposed to be a comma-separated
list.

However… `debug.mono.log` *itself* uses commas to separate components,
so using the following:

	$ adb shell setprop debug.mono.log mono_log_mask=asm,gc,dll

would *not* set the `$MONO_LOG_MASK` environment variable to the
expected value of `asm,gc,dll`, but would instead it would end up being
just `asm`, as the other values would be considered to be
`debug.mono.log` components instead.

To fix this, change the `mono_log_mask` separator to `:`, so that
we would now use:

	$ adb shell setprop debug.mono.log mono_log_mask=asm:gc:dll

Additionally, document all the `debug.mono.*` properties and remove
the `DEBUG_MONO_GDBPORT_PROPERTY` constant, since it's not used
anywhere.

Also, when `assembly` or `gc` logging is enabled, we now set the Mono
tracing level to `info` instead of the default `error`.  This makes
tracking assembly loader and garbage collector issues easier by
showing more info from the Mono runtime, not just from Xamarin.Android.