[Xamarin.Android.Build.Tests] Update Microsoft_Azure_EventHubs
We are seeing a Component Governance error for `System.Net.WebSockets.Client`.
This is because one of our tests uses `Microsoft.Azure.EventHubs`. This
is pulling in an old version of `System.Net.WebSockets.Client`.

So we need to upgrade the NuGet to make sure we pull in the newer version.
Interestingly, with the newer version we no longer get the expected
failure when using `PackageReference`. This is probably due to them fixing
the NuGet. So we have to update the test to reflect that change.