[Xamarin.Android.Build.Tests] Update Microsoft_Azure_EventHubs (#6519)
Fixes: https://devdiv.visualstudio.com/DevDiv/_componentGovernance/112013/alert/2194574?typeId=6317076

We are seeing a Component Governance error for
`System.Net.WebSockets.Client`.  This is because one of our tests use
`Microsoft.Azure.EventHubs`, which is pulling in an old version of
`System.Net.WebSockets.Client`.

We need to upgrade the NuGet to make sure we pull in a newer version.
Interestingly, with the newer version we no longer get the expected
failure when using `@(PackageReference)`, probably due to the NuGet
package being fixed.

Update the test to reflect that change.