[WIP] Hash names of all the packaged DSOs
Use the hash to map any form of DSO name sent to us by Mono (we generate
hashes for several mutations) to find a cache entry for that DSO. Cache
includes the actual DSO name, whether it should be ignored when
requested (true for empty AOT DSOs) and the cached module handle.

This removes a bit of string processing from both startup and the
application run time.

Timing TBD