[ci] Move to 1ES Windows VS 2022 pool
Context: https://aka.ms/1eshostedpools

I've created a new Windows VS 2022 pool to use for Windows build and
test jobs.  The azure resource for this pool can be [found here][0].

The base image of this pool is the Microsoft-hosted [windows-2022][1]
image, and our machines will stay in sync with any updates that happen
to that hosted pool moving foward.

[0]: https://ms.portal.azure.com/#@microsoft.onmicrosoft.com/resource/subscriptions/cd4829e2-e38b-43d2-8316-2f2009f36f97/resourceGroups/devdiv-vm-scale-sets/providers/Microsoft.CloudTest/hostedpools/android-win-2022/overview
[1]: https://github.com/actions/virtual-environments/blob/ddaeaaa8fba88b009ab140001e63ee317dae52db/images/win/Windows2022-Readme.md