[generator] Use the correct javadoc element (#934)
Fixes: https://github.com/xamarin/java.interop/issues/933

Fixes a bug when attempting to find the `<javadoc/>` element that is
associated with a specific member name and jni-signature.  When a type
contained multiple members with the same name, we would always return
the first `<javadoc/>` that matched the member name, rather than the
one that matched both the name and signature.